requirements.txt
matplotlib==3.5.1
numpy==1.21.2
pandas==1.2.5
Pillow==8.4.0
plotly==5.4.0
scikit_learn==1.0.1
seaborn==0.11.2
squarify==0.4.3
streamlit==1.2.0
